ensemble graph component (#9481)
* ensemble graph component

* minor

* minor fix (import from ensemble not _ensemble)

* fix (revert _ensemble -_-); change (re-introduce Prediction; add Prediction2)

* fix

* fix type

* fix (revert _ensemble)

* fix (policy prediction -> 2)

* revert max confidence vs prediction

* revert max predicton -> confidence

* adapted some TODOs; made _best* funtions static; arg_max; contains_no_form_predictions

* Apply suggestions from code review


* lint

* minor (add expected prefix in test and rearrange a bit)

* small fix (static -> class; revert validate against domain)

* Apply suggestions from code review


* revert PolicyPrediction (confidence, priority is tuple); commented out validate_against_domain which needs to be adapted in rulepolicy

* minor (warn_bla -> warn_if_bla)

* rm comments/todos

* compare using priority-tuple

* remove validate_* and stuff moved to new components

* rm comment on action metadata

* Apply suggestions from code review


* revert to if-version

* revert (accidentally removed _get_featurizer_from_config)

* Update tests/core/test_ensemble.py


* lint; rm comments; rm unused imports

* add (old is_not_in_training_data - with new/fixed docstr, tests for empty prediction list, interface with kwargs), rm (duplicated test), change (all tests use kwargs version)

* typo

* fix (docstr)
